Quizz and formations initialize with data folders

+++ formation/src/main/java/com/learnandphish/formation/dto/FormationRequest.java
@@ -1,4 +1,4 @@
 package com.learnandphish.formation.dto;
 
-public record FormationRequest(Long id, String name, String description) {
+public record FormationRequest(Integer id, String name, String description) {
 }
