+++ formation/src/main/java/com/learnandphish/formation/repository/FormationRepository.java
@@ -3,5 +3,5 @@ package com.learnandphish.formation.repository;
 import com.learnandphish.formation.model.Formation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FormationRepository extends JpaRepository<Formation, Long> {
+public interface FormationRepository extends JpaRepository<Formation, Integer> {
 }
